+++ src/rest/streaming.rs
@@ -1,5 +1,4 @@
 mod batch;
-mod capture_messages;
 mod forward_to_file;
 mod forward_to_std;
 mod messaging;
